Modify default page on store page
